Add Roles to control users access to api create migrations to add roles and users_roles table refactor userdetails to be an adapter class create the class roleModel as an Entity add roles to UserModel implement to RoleModel the GrantedAuthority interface and return the authorities in the adpter class Userdetails
