+++ src/main/resources/db/migration/V5__create_table_role.sql
@@ -1,0 +1,5 @@
+CREATE TABLE roles(
+`id` BIGINT NOT NULL AUTO_INCREMENT,
+`name` VARCHAR(50) NOT NULL,
+PRIMARY KEY(`id`)
+);
